# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-tenant SaaS POS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches
      parameters:
        - { name: organizationId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Users
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List product categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Products
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /customers/{id}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      summary: Get customer order history
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # Orders / Transactions
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: string }

    # Auth
    RegisterInput:
      type: object
      required: [name, email, password, organizationName]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        organizationName: { type: string }
        businessType:
          type: string
          enum: [retail, grocery, pharmacy, restaurant, cafe, salon, service, hardware, convenience, wholesale, boutique, other]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token: { type: string }
    UserProfile:
      type: object
      required: [id, name, email, role, organizationId]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        organizationId: { type: integer }
        branchId: { type: ["integer", "null"] }
        organizationName: { type: string }
        avatar: { type: ["string", "null"] }

    # Organization
...
    Branch:
      type: object
      required: [id, organizationId, name, status, createdAt]
      properties:
        id: { type: integer }
        organizationId: { type: integer }
        name: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        status: { type: string, enum: [active, inactive] }
        isMain: { type: boolean }
        taxRate: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
    BranchInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        isMain: { type: boolean }
        taxRate: { type: number }
    BranchUpdate:
      type: object
      properties:
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        status: { type: string }
        taxRate: { type: number }

    # User
...
    Category:
      type: object
      required: [id, name, productCount]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        icon: { type: ["string", "null"] }
        productCount: { type: integer }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        color: { type: string }
        icon: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        color: { type: string }
        icon: { type: string }

    # Product
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        loyaltyPoints: { type: integer }
        totalSpent: { type: number }
        orderCount: { type: integer }
        notes: { type: ["string", "null"] }
        birthday: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        birthday: { type: string }
        notes: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        birthday: { type: string }
        notes: { type: string }

    # Order
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        discount: { type: number }
        totalPrice: { type: number }
        notes: { type: ["string", "null"] }
    Order:
      type: object
      required: [id, orderNumber, branchId, status, subtotal, total, createdAt]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        branchId: { type: integer }
        branchName: { type: string }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        cashierId: { type: integer }
        cashierName: { type: string }
        status: { type: string, enum: [draft, held, completed, refunded, voided] }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxAmount: { type: number }
        total: { type: number }
        paymentMethod: { type: ["string", "null"] }
        amountPaid: { type: ["number", "null"] }
        change: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"] }
...